vnc-server: Use int type to read port numbers
The `GrdSettingsClass` properties to store port numbers are type of `int`.
However, the parameters of the `g_object_get` function point to the
variables of the `uint16_t` type. This somehow work on little-endian
systems, but not on big-endian. Let's read the port numbers to variable of
the `int` type to fix port handling on big-endian systems.

+++ src/grd-vnc-server.c
@@ -135,7 +135,7 @@ grd_vnc_server_start (GrdVncServer  *vnc_server,
                       GError       **error)
 {
   GrdSettings *settings = grd_context_get_settings (vnc_server->context);
-  uint16_t vnc_port;
+  int vnc_port;
   uint16_t selected_vnc_port = 0;
   gboolean negotiate_port;
   GrdDBusRemoteDesktopVncServer *vnc_server_iface;
